README.md: Move commands to code blocks
Commands in code blocks are easier to read, because they are clearly
standing apart from the body of the text and don't risk being wrapped by
accident. They are also easier to copy-paste, because they're easier to
select or, even better, because GitHub adds a copy button to code
blocks (but not to inline literals).

Signed-off-by: Quentin Monnet <quentin@isovalent.com>